+++ src/onegov.user/onegov/user/upgrade.py
@@ -5,7 +5,7 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 from onegov.core.upgrade import upgrade_task
 from onegov.core.orm.types import JSON
 from onegov.user import User
-from sqlalchemy import Boolean, Column
+from sqlalchemy import Boolean, Column, Text
 
 
 @upgrade_task('Add second_factor column')
@@ -26,3 +26,14 @@ def add_active_column(context):
 
     context.session.flush()
     context.operations.alter_column('users', 'active', nullable=False)
+
+
+@upgrade_task('Add realname column')
+def add_realname_column(context):
+    context.operations.add_column(
+        'users', Column('realname', Text, nullable=True))
+
+    for user in context.session.query(User).all():
+        user.realname = (user.data and user.data or {}).get('name')
+
+    context.session.flush()